server/requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
aiofiles
pyinstaller
